ws: removes bufferpool as internal dep
